Add event to environment service for navigation
Added a new event on the environment service for triggering a navigation
of the host environment window.
Added documentation.
Renamed the ‘send’ event method on a Card object to ‘trigger’.
Bumped the SDK version for the first time.  Going to continue doing 
minor version bumps when the client SDK must be changed, until we hit 
the 1.0 release, when we will begin following semver.