Move away from using consumer object for activate
Trying to fix an issue where the activate event is never being handled.
It appears that the connection with the hash including the Environment 
Consumer is happening asynchronously, which means that a developer can 
call `card.services(‘environment’)` before the Environment Consumer 
handles the `’activate’` event.  It looks like that call, which connects 
to the `’environment’` port is causing the `’activate’` event to be 
processed before the Consumer is connected to handle it.
So, instead of setting up a Consumer, just connect to the environment 
service immediately and handle the activate event using the `’services’`
method so that the port connection can be created and cached for future 
calls.
Added the ability to pass in a callback context to `CardService#on`.